+++ src/global-bonuses/specific/bonuses/spirited-charge-bonus.mjs
@@ -2,18 +2,18 @@ import { SpecificBonus } from '../../../bonuses/_specific-bonus.mjs';
 import { showEnabledLabel } from '../../../handlebars-handlers/enabled-label.mjs';
 import { LanguageSettings } from '../../../util/settings.mjs';
 
-export class PreciseShot extends SpecificBonus {
+export class SpiritedCharge extends SpecificBonus {
     /** @inheritdoc @override */
-    static get sourceKey() { return 'precise-shot'; }
+    static get sourceKey() { return 'spirited-charge'; }
 
     /** @inheritdoc @override */
-    static get journal() { return 'Compendium.ckl-roll-bonuses.roll-bonuses-documentation.JournalEntry.FrG2K3YAM1jdSxcC.JournalEntryPage.4A4bCh8VsQVbTsAY#precise-shot'; }
+    static get journal() { return 'Compendium.ckl-roll-bonuses.roll-bonuses-documentation.JournalEntry.FrG2K3YAM1jdSxcC.JournalEntryPage.4A4bCh8VsQVbTsAY#spirited-charge'; }
 
     /** @inheritdoc @override @returns {CreateAndRender} */
     static get configuration() {
         return {
             type: 'render-and-create',
-            compendiumId: '53urYIbYYpQuoSLd',
+            compendiumId: 'jKVhOxHu6VmUgFdp',
             isItemMatchFunc: (name) => name === Settings.name,
             showInputsFunc: (item, html, isEditable) => {
                 showEnabledLabel({
@@ -31,9 +31,9 @@ export class PreciseShot extends SpecificBonus {
 }
 
 class Settings {
-    static get name() { return LanguageSettings.getTranslation(PreciseShot.key); }
+    static get name() { return LanguageSettings.getTranslation(SpiritedCharge.key); }
 
     static {
-        LanguageSettings.registerItemNameTranslation(PreciseShot.key);
+        LanguageSettings.registerItemNameTranslation(SpiritedCharge.key);
     }
 }