+++ types/roll-bonus-api.d.ts
@@ -173,7 +173,7 @@ import {
     intersection,
     intersects,
 } from '../src/util/array-intersects.mjs';
-import { addCheckToAttackDialog } from '../src/util/attack-dialog-helper.mjs';
+import { addCheckToAttackDialog, addTextInputToAttackDialog, getFormData } from '../src/util/attack-dialog-helper.mjs';
 import { confirmationDialog } from '../src/util/confirmation-dialog.mjs';
 import { getEnhancementBonusForAction } from '../src/util/enhancement-bonus-helper.mjs';
 import { FormulaCacheHelper, getDocFlags } from '../src/util/flag-helpers.mjs';
@@ -208,6 +208,7 @@ import { signed } from '../src/util/to-signed-string.mjs';
 import { Trait } from '../src/util/trait-builder.mjs';
 import { truthiness } from '../src/util/truthiness.mjs';
 import { distinct, uniqueArray } from '../src/util/unique-array.mjs';
+import { SpiritedCharge } from '../src/global-bonuses/specific/bonuses/spirited-charge-bonus.mjs';
 
 export class _RollBonusesAPI {
     es: any;
@@ -387,8 +388,10 @@ export class _RollBonusesAPI {
         ['weapon-specialization']: typeof WeaponSpecialization;
         ['weapon-specialization-greater']: typeof WeaponSpecializationGreater;
 
+        /** specific bonuses gratned from enabling global bonuses  */
         ['outflank']?: typeof Outflank;
         ['precise-shot']?: typeof PreciseShot;
+        ['spirited-charge']?: typeof SpiritedCharge;
     };
     /** Array of all targeted targets */
     get allSpecificBonusTypes(): (typeof SpecificBonus)[];
@@ -466,11 +469,13 @@ export class _RollBonusesAPI {
         };
 
         addCheckToAttackDialog: typeof addCheckToAttackDialog;
+        addTextInputToAttackDialog: typeof addTextInputToAttackDialog;
         confirmationDialog: typeof confirmationDialog;
         currentTargetedActors: typeof currentTargetedActors;
         currentTargets: typeof currentTargets;
         getActionDamageTypes: typeof getActionDamageTypes;
         getDocFlags: typeof getDocFlags;
+        getFormData: typeof getFormData;
         getEnhancementBonusForAction: typeof getEnhancementBonusForAction;
         getSkillFormula: typeof getSkillFormula;
         getSourceFlag: typeof getSourceFlag;